+++ .gitignore
@@ -1,3 +1,4 @@
 # testing directory
 testing/
 testing_import/
+SASA_example/
