+++ build_recipe/meta.yaml
@@ -3,7 +3,7 @@ package:
     version: 0.1
 
 source:
-    path : ..
+    path : ../src
 
 requirements:
     build:
@@ -11,7 +11,7 @@ requirements:
 
     run:
         - python
-        - numpy
+        - numpy==1.26.0
         - ovito==3.9.2
         - vmd-python==3.1.4
