+++ main.py
@@ -1,0 +1,3 @@
+from sasa_lammps import sasa
+
+print(sasa.__doc__)
