+++ sasa_lammps/__init__.py
@@ -1,0 +1,1 @@
+from sasa_lammps.sasa_lammps import sasa
